requirements.txt
streamlit
praw
pandas
python-dotenv